src/MyRecipes.Application/UseCases/User/Profile/GetUserProfileUseCase.cs
using AutoMapper;
using MyRecipes.Communication.Responses;
using MyRecipes.Domain.Services.LoggedUser;

namespace MyRecipes.Application.UseCases.User.Profile;
internal class GetUserProfileUseCase : IGetUserProfileUseCase
{
    private readonly ILoggedUser _loggedUser;
    private readonly IMapper _mapper;

    public GetUserProfileUseCase(ILoggedUser loggedUser, IMapper mapper)
    {
        _loggedUser = loggedUser;
        _mapper = mapper;
    }

    public async Task<ResponseUserProfileJson> Execute()
    {
        var user = await _loggedUser.User();

        return _mapper.Map<ResponseUserProfileJson>(user);
    }
}
